hitID -1/15 isn't special in the way previously thought
it's special in another way, you can hit several in a row, just not multiple
frames of the same hitbox (which is probably always true)

encode hitbox index for now, the real question is whether the same hitbox ID
after inheriting ID would hit again? probably would unless the game does it
like that too

+++ guy.hpp
@@ -572,7 +572,7 @@ private:
     std::deque<int> directionBuffer;
 
     // hitting side
-    int canHitID = 0;
+    uint64_t canHitID = 0;
     bool hitThisFrame = false;
     bool hasBeenBlockedThisFrame = false;
     bool hitArmorThisFrame = false;
